Improved error reporting support.
Notably, only report the short exception message for common and known
exceptions (IOError and new AudioFormatError), but full tracebacks for all other
exceptions in all cases.
Also, as noted, if Mutagen fails to open a file properly (because it's not what
its extension purports to be), an AudioFormatError is raised, which gets printed
in short form.